reorganizing example, helper functions, and model code
